+++ src/com/gisfederal/request/FilterThenHistogramRequest.java
@@ -7,9 +7,7 @@ import java.util.List;
 
 import org.apache.log4j.Logger;
 
-import avro.java.gpudb.add_object_request;
 import avro.java.gpudb.filter_then_histogram_request;
-import avro.java.gpudb.histogram_request;
 
 import com.gisfederal.AvroUtils;
 import com.gisfederal.GPUdb;
@@ -21,15 +19,17 @@ import com.gisfederal.SetId;
  */
 public class FilterThenHistogramRequest extends Request {
 
-	public FilterThenHistogramRequest(GPUdb gPUdb, String file, SetId id, String filter_attribute, List<CharSequence> filter, String histogram_attribute, long interval, double start, double end) {
+	public FilterThenHistogramRequest(GPUdb gPUdb, String file, SetId id, String filter_attribute, List<CharSequence> filter, 
+			String histogram_attribute, double interval, double start, double end) {
 		this.gPUdb = gPUdb;
 		this.file = file;
 		this.log = Logger.getLogger(FilterThenHistogramRequest.class);
 		
 		log.debug("file:"+file+" id:"+id.get_id()+" start:"+start+" end:"+end);
-		filter_then_histogram_request request = new filter_then_histogram_request(histogram_attribute, end, interval, id.get_id(), start, filter_attribute, filter, this.gPUdb.getUserAuth());
+		filter_then_histogram_request request = new filter_then_histogram_request(histogram_attribute, end, interval, id.get_id(), start, 
+				filter_attribute, filter, this.gPUdb.getUserAuth());
 		this.requestData = new RequestData(AvroUtils.convert_to_bytes(request));
-
+		
 		// Create log msg for audit
 		createAuditMsg(request);
 	}
